Moved parseAchievementName(String[]) to AbstractCommand
This method is useful outside of AbstractParsableCommand in upcoming inspect command.
