+++ advanced-achievements-plugin/src/main/java/com/hm/achievement/command/executable/AbstractCommand.java
@@ -1,5 +1,6 @@
 package com.hm.achievement.command.executable;
 
+import org.apache.commons.lang3.StringUtils;
 import org.bukkit.ChatColor;
 import org.bukkit.Sound;
 import org.bukkit.command.CommandSender;
@@ -80,5 +81,23 @@ public abstract class AbstractCommand implements Reloadable {
 			sound = Sound.ENTITY_FIREWORK_ROCKET_BLAST;
 		}
 		player.getWorld().playSound(player.getLocation(), sound, 1, 0.7f);
+	}
+
+	/**
+	 * Extracts the name of the achievement from the command line arguments.
+	 *
+	 * @param args
+	 * @return the achievement name
+	 */
+	String parseAchievementName(String[] args) {
+		StringBuilder achievementName = new StringBuilder();
+		// Rebuild name of achievement by concatenating elements in the string array. The name of the player is last.
+		for (int i = 1; i < args.length - 1; i++) {
+			achievementName.append(args[i]);
+			if (i != args.length - 2) {
+				achievementName.append(' ');
+			}
+		}
+		return StringUtils.replace(achievementName.toString(), "\u2423", " ");
 	}
 }
